Move Labels to Conference
