+++ react/features/conference/components/AbstractLabels.js
