+++ react/features/large-video/components/styles.js
@@ -1,7 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { BoxModel, ColorPalette, createStyleSheet } from '../../base/styles';
-import { FILMSTRIP_SIZE } from '../../filmstrip';
+import { ColorPalette, createStyleSheet } from '../../base/styles';
 
 /**
  * Size for the Avatar.
@@ -9,34 +8,6 @@ import { FILMSTRIP_SIZE } from '../../filmstrip';
 export const AVATAR_SIZE = 200;
 
 export default createStyleSheet({
-
-    /**
-     * View that contains the indicators.
-     */
-    indicatorContainer: {
-        flex: 1,
-        flexDirection: 'row',
-        justifyContent: 'flex-end',
-        margin: BoxModel.margin
-    },
-
-    /**
-     * Indicator container for wide aspect ratio.
-     */
-    indicatorContainerWide: {
-        marginRight: FILMSTRIP_SIZE + BoxModel.margin
-    },
-
-    labelWrapper: {
-        flexDirection: 'column',
-        position: 'absolute',
-        right: 0,
-
-        // Both on Android and iOS there is the status bar which may be visible.
-        // On iPhone X there is the notch. In the two cases BoxModel.margin is
-        // not enough.
-        top: BoxModel.margin * 3
-    },
 
     /**
      * Large video container style.
